requirements.txt
Django==5.2
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
psycopg2==2.9.9
django-cors-headers==4.7.0
social-auth-app-django==5.2.0
python-social-auth==0.3.6
requests==2.31.0
dj-rest-auth==7.0.1
django-allauth==65.7.0